+++ smawg/__init__.py
@@ -4,9 +4,9 @@ See https://github.com/expurple/smawg for more info about the project.
 """
 
 from smawg._engine import Game, Hooks, InvalidAssets, validate
-from smawg._plugin_interface import Ability, Combo, Player, Race, Region
-from smawg.exceptions import RulesViolation
-
+from smawg._plugin_interface import (
+    Ability, Combo, Player, Race, Region, RulesViolation
+)
 
 __all__ = [
     "Game", "Hooks", "InvalidAssets", "validate",